+++ Engine2/Simulation Runtime/Engine/System/Position/Component/CPosition.swift
@@ -2,6 +2,7 @@
 ///
 /// The value is stored in `World.positionComponents`; positioned entity
 /// facades read this row live, and simulation systems mutate the store directly.
+/// Coordinates use double-precision meters.
 struct CPosition: PComponent {
-    var position: SIMD3<Float>
+    var position: SIMD3<Double>
 }
